tools/org_cleaner/requirements.txt
click
google-cloud-asset
google-cloud-compute
google-cloud-logging
google-cloud-org-policy
google-cloud-resource-manager